Library of plotting functions
These functions are needed together with the complementary R script to create the exact same box plots and dimension plots found in the paper.
The plotting utilities are implemented in the FlowR bundle, as well.
Please note that the code is in need of refactoring. In the beginning of putting this library together I've tried to avoid using the factor type, which overcomplicates a few things.